# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
  domain.Client:
    properties:
      address:
        type: string
      city:
        type: string
      createdAt:
        type: string
      dateOfBirth:
        type: string
      deletedAt:
        description: Soft delete
        type: string
      dni:
        description: Spanish ID (DNI/NIE/NIF)
        type: string
      email:
        type: string
      firstName:
        type: string
      id:
        type: string
      isActive:
        type: boolean
      lastName:
        type: string
      lastVisit:
        type: string
      notes:
        type: string
      phone:
        type: string
      postalCode:
        type: string
      province:
        type: string
      updatedAt:
        type: string
      userId:
        description: 'Optional: if client has login access'
        type: string
    type: object
  domain.User:
    properties:
      createdAt:
        type: string
      email:
        type: string
      firstName:
        type: string
      id:
        type: string
      isActive:
        type: boolean
      lastName:
        type: string
      role:
        $ref: '#/definitions/domain.UserRole'
      updatedAt:
        type: string
    type: object
  domain.UserRole:
    enum:
    - admin
    - employee
    - client
    type: string
    x-enum-varnames:
    - RoleAdmin
    - RoleEmployee
    - RoleClient
  service.AuthResponse:
    properties:
      token:
        type: string
      user:
        $ref: '#/definitions/domain.User'
    type: object
  service.ClientListResponse:
    properties:
      clients:
        items:
          $ref: '#/definitions/domain.Client'
        type: array
      page:
        type: integer
      pageSize:
        type: integer
      total:
        type: integer
      totalPages:
        type: integer
    type: object
  service.CreateClientRequest:
    properties:
      address:
        type: string
      city:
        type: string
      dateOfBirth:
        description: 'Format: YYYY-MM-DD'
        type: string
      dni:
        type: string
      email:
        type: string
      firstName:
        type: string
      lastName:
        type: string
      notes:
        type: string
      phone:
        type: string
      postalCode:
        type: string
      province:
        type: string
      userId:
        type: string
    required:
    - dni
    - email
    - firstName
    - lastName
    - phone
    type: object
  service.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  service.RegisterRequest:
    properties:
      email:
        type: string
      firstName:
        type: string
      lastName:
        type: string
      password:
        minLength: 8
        type: string
      role:
        type: string
    required:
    - email
    - firstName
    - lastName
    - password
    type: object
  service.UpdateClientRequest:
    properties:
      address:
        type: string
      city:
        type: string
      dateOfBirth:
        description: 'Format: YYYY-MM-DD'
        type: string
      dni:
        type: string
      email:
        type: string
      firstName:
        type: string
      isActive:
        type: boolean
      lastName:
        type: string
      notes:
        type: string
      phone:
        type: string
      postalCode:
        type: string
      province:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@arnela.com
    name: API Support
  description: API for Arnela professional office management system
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Arnela CRM/CMS API
  version: "1.0"
paths:
  /api/v1/auth/login:
    post:
      consumes:
      - application/json
      description: Authenticates a user and returns a JWT token
      parameters:
      - description: Login credentials
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/service.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/service.AuthResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Login user
      tags:
      - auth
  /api/v1/auth/me:
    get:
      description: Returns the authenticated user's information
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.User'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get current user
      tags:
      - auth
  /api/v1/auth/register:
    post:
      consumes:
      - application/json
      description: Creates a new user account
      parameters:
      - description: Registration data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/service.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/service.AuthResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Register a new user
      tags:
      - auth
  /api/v1/clients:
    get:
      description: Retrieves a paginated list of clients with optional filters (admin/employee
        only)
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 20
        description: Page size
        in: query
        name: pageSize
        type: integer
      - description: Search in name, email, phone, DNI
        in: query
        name: search
        type: string
      - description: Filter by active status
        in: query
        name: isActive
        type: boolean
      - description: Filter by city
        in: query
        name: city
        type: string
      - description: Filter by province
        in: query
        name: province
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/service.ClientListResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: List clients
      tags:
      - clients
    post:
      consumes:
      - application/json
      description: Creates a new client (admin/employee only)
      parameters:
      - description: Client data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/service.CreateClientRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/domain.Client'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Conflict
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create a new client
      tags:
      - clients
  /api/v1/clients/{id}:
    delete:
      description: Soft-deletes a client (admin only)
      parameters:
      - description: Client ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete client
      tags:
      - clients
    get:
      description: Retrieves a client's information (admin/employee only)
      parameters:
      - description: Client ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Client'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get client by ID
      tags:
      - clients
    put:
      consumes:
      - application/json
      description: Updates a client's information (admin/employee only)
      parameters:
      - description: Client ID
        in: path
        name: id
        required: true
        type: string
      - description: Client update data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/service.UpdateClientRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Client'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Conflict
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update client
      tags:
      - clients
  /api/v1/clients/me:
    get:
      description: Retrieves the authenticated client's information (client role only)
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Client'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get my client info
      tags:
      - clients
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"